+++ changelog.md
@@ -1,5 +1,6 @@
 ##### Version 1.16.1-4.0.0-pre
 - Re-added custom banner patterns. Craft the blueprint, make patterns, then use them on the loom! (BluSunrize)
+- Added a warning when switching to "fabulous" graphics while stencil is enabled (Malte)
 - Fixed log spam caused by wire sync issues (Malte)
 - Fixed manual page for shaders (BluSunrize)